% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.plasso.R
\name{add_intercept}
\alias{add_intercept}
\title{Adds an intercept to a matrix}
\usage{
add_intercept(mat)
}
\arguments{
\item{mat}{Any matrix (with column names).}
}
\value{
Matrix with intercept.
}
\description{
\code{\link{add_intercept}} adds an intercept to a matrix.
}
\keyword{internal}
