% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasso.R
\name{coef.plasso}
\alias{coef.plasso}
\title{Extract coefficients from a plasso object}
\usage{
\method{coef}{plasso}(object, ..., s = NULL)
}
\arguments{
\item{object}{\code{\link[plasso]{plasso}} object}

\item{...}{Pass generic \code{\link[stats]{coef}} options}

\item{s}{If Null, coefficients are returned for all lambda values. If a value is provided, the closest lambda value of the plasso object is used.}
}
\value{
List containing matrices or vector of coefficients for both Lasso and Post-Lasso.
}
\description{
Extract coefficients for both Lasso and Post-Lasso from a plasso object.
}
