% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.plasso.R
\name{cv.plasso}
\alias{cv.plasso}
\title{Cross-validated Lasso and Post-Lasso}
\usage{
cv.plasso(x, y, w = NULL, kf = 10, parallel = FALSE, ...)
}
\arguments{
\item{x}{Matrix of covariates (number of observations times number of covariates matrix)}

\item{y}{Vector of outcomes}

\item{w}{Vector of weights}

\item{kf}{Number of folds in k-fold CV}

\item{parallel}{Set as TRUE for parallelized cross-validation. Default is FALSE.}

\item{...}{Pass \code{\link[glmnet]{glmnet}} options}
}
\value{
List object including base \code{\link[glmnet]{glmnet}} object and cross-validation results (incl. optimal Lambda values) for both Lasso and Post-Lasso model.
\item{call}{the call that produced this}
\item{lasso_full}{base \code{\link[glmnet]{glmnet}} object}
\item{kf}{number of folds in k-fold CV}
\item{cv_MSE_lasso}{cross-validated MSEs of Lasso model (for every kth iteration of k-fold cross-validation)}
\item{cv_MSE_plasso}{cross-validated MSEs of Post-Lasso model (for every kth iteration of k-fold cross-validation)}
\item{mean_MSE_lasso}{averaged cross-validated MSEs of Lasso model}
\item{mean_MSE_plasso}{averaged cross-validated MSEs of Post-Lasso model}
\item{ind_min_l}{index of MSE optimal lambda value for Lasso model}
\item{ind_min_pl}{index of MSE optimal lambda value for Post-Lasso model}
\item{lambda_min_l}{MSE optimal lambda value for Lasso model}
\item{lambda_min_pl}{MSE optimal lambda value for Post-Lasso model}
\item{names_l}{Names of active variables for MSE optimal Lasso model}
\item{names_pl}{Names of active variables for MSE optimal Post-Lasso model}
\item{coef_min_l}{Coefficients for MSE optimal Lasso model}
\item{coef_min_pl}{Coefficients for MSE optimal Post-Lasso model}
\item{x}{Input matrix of covariates}
\item{y}{Matrix of outcomes}
\item{w}{Matrix of weights}
}
\description{
\emph{cv.plasso()} uses the \code{\link[glmnet]{glmnet}} package to estimate the coefficient paths and cross-validates least squares Lasso AND Post-Lasso.
}
\examples{
data(toeplitz)
y = as.matrix(toeplitz[,1])
X = toeplitz[,-1]
\donttest{p.cv = plasso::cv.plasso(X,y)}
\donttest{print(summary(p.cv, default=FALSE))}
\donttest{plot(p.cv, legend_pos="left")}

}
