% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.plasso.R
\name{CV_core}
\alias{CV_core}
\title{Core part for (Post-) Lasso cross-validation}
\usage{
CV_core(x, y, w, cvgroup, list, i, lambda, ...)
}
\arguments{
\item{x}{Covariate matrix to be used in cross-validation}

\item{y}{Vector of outcomes}

\item{w}{Vector of weight}

\item{cvgroup}{Categorical with k groups to identify folds}

\item{list}{List 1:k}

\item{i}{Number of fold that is used for prediction}

\item{lambda}{Series of lambdas used}

\item{...}{Pass \code{\link[glmnet]{glmnet}} options}
}
\value{
MSE_lasso / MSE_plasso: means squared errors for each lambda.
}
\description{
\code{\link{CV_core}} contains the core parts of the cross-validation for Lasso and Post-Lasso.
}
\keyword{internal}
