% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.plasso.R
\name{find_Xse_ind}
\alias{find_Xse_ind}
\title{Helper function to find the position for prespecified SE rules}
\usage{
find_Xse_ind(CV, ind_min, oneSE, factor)
}
\arguments{
\item{CV}{Vector of cross-validated criterion}

\item{ind_min}{Index of cross-validated minimum}

\item{oneSE}{Vector that contains the standard errors of the cross-validated criterion for the whole grid}

\item{factor}{Factor in which direction to go: Negative values favor smaller models, positive values favor larger models}
}
\value{
Index on the Lambda grid.
}
\description{
\code{\link{find_Xse_ind}} is a helper function that finds the position for prespecified SE rules.
}
\keyword{internal}
