% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.plasso.R
\name{fit_betas}
\alias{fit_betas}
\title{plasso fitting}
\usage{
fit_betas(x, y, w, nm_act, coef_lasso)
}
\arguments{
\item{x}{Matrix of covariates (number of observations times number of covariates matrix)}

\item{y}{Vector of outcomes}

\item{w}{Vector of weights}

\item{nm_act}{Vector of active variables}

\item{coef_lasso}{Vector of lasso coefficients}
}
\value{
Beta estimates.
}
\description{
\code{\link{fit_betas}} estimates OLS model only for active coefficients (from lasso)
}
\keyword{internal}
