% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasso.R
\name{fitted_values}
\alias{fitted_values}
\title{Fitted values for a subset of active variables}
\usage{
fitted_values(XtX_all, Xty_all, x_pred, nm_act)
}
\arguments{
\item{XtX_all}{Crossproduct of all covariates}

\item{Xty_all}{Crossproduct of covariates and outcome}

\item{x_pred}{Covariates matrix of the PREDICTION sample}

\item{nm_act}{Names of active variables}
}
\value{
Fitted values in the prediction sample.
}
\description{
\emph{fitted_values()} extracts a subset of active variables (nm_act) of the
relevant variables from X'X and X'y to get out-of-sample predictions
for a matrix containing only the active variables.
This speeds up the cross-validation for Post-Lasso to a large extent.
}
\keyword{internal}
