% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.plasso.R
\name{norm_w_to_n}
\alias{norm_w_to_n}
\title{Normalization of sample weights for potential sample weights}
\usage{
norm_w_to_n(w, d = NULL)
}
\arguments{
\item{w}{Vector or n x 1 matrix of weights that should be normalized}

\item{d}{Vector of treatment indicators}
}
\value{
Normalized weights.
}
\description{
\code{\link{norm_w_to_n}} normalizes weights either to N or to N in treated and controls separately.
}
\keyword{internal}
