% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.plasso.R
\name{print.cv.plasso}
\alias{print.cv.plasso}
\title{Print (Post-) Lasso model}
\usage{
\method{print}{cv.plasso}(x, ..., digits = max(3, getOption("digits") - 3))
}
\arguments{
\item{x}{\code{\link[plasso]{cv.plasso}} object}

\item{...}{Pass generic \code{\link[base]{print}} options}

\item{digits}{Integer, used for number formatting}
}
\value{
Prints cross-validated MSE and active variables for Lasso and Post-Lasso.
}
\description{
Printing main insights from (Post-) Lasso model.
}
