% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasso.R
\name{print.plasso}
\alias{print.plasso}
\title{Print (Post-) Lasso model}
\usage{
\method{print}{plasso}(x, ..., digits = max(3, getOption("digits") - 3))
}
\arguments{
\item{x}{\code{\link{plasso}} object}

\item{...}{Pass generic \code{\link[base]{print}} options}

\item{digits}{Integer, used for number formatting}
}
\value{
Prints \code{\link[glmnet]{glmnet}}-like output.
}
\description{
Printing main insights from (Post-) Lasso model.
}
