% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.plasso.R
\name{print.summary.cv.plasso}
\alias{print.summary.cv.plasso}
\title{Print summary of (Post-) Lasso model}
\usage{
\method{print}{summary.cv.plasso}(x, ..., digits = max(3L, getOption("digits") - 3L))
}
\arguments{
\item{x}{Summary of plasso object (either of class \code{\link{summary.cv.plasso}} or \code{\link[base]{summary}})}

\item{...}{Pass generic R \code{\link[base]{print}} options}

\item{digits}{Integer, used for number formatting}
}
\value{
Prints information from \code{\link{summary.cv.plasso}} object into console.
}
\description{
Prints summary information of \code{\link{cv.plasso}} object
}
