% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasso.R
\name{summary.plasso}
\alias{summary.plasso}
\title{Summary of (Post-) Lasso model}
\usage{
\method{summary}{plasso}(object, ...)
}
\arguments{
\item{object}{\code{\link[plasso]{plasso}} object}

\item{...}{Pass generic \code{\link[base]{summary}} summary options}
}
\value{
'summaryDefault' object
}
\description{
Summary of (Post-) Lasso model.
}
