% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.plasso.R
\name{cv.plasso}
\alias{cv.plasso}
\title{Cross-validated Lasso and Post-Lasso}
\usage{
cv.plasso(x, y, w = NULL, kf = 10, parallel = FALSE, ...)
}
\arguments{
\item{x}{Matrix of covariates (number of observations times number of covariates matrix)}

\item{y}{Vector of outcomes}

\item{w}{Vector of weights}

\item{kf}{Number of folds in k-fold CV}

\item{parallel}{Set as TRUE for parallelized cross-validation. Default is FALSE.}

\item{...}{Pass \code{\link[glmnet]{glmnet}} options}
}
\value{
List object including base \code{\link[glmnet]{glmnet}} object and cross-validation results (incl. optimal Lambda values) for both Lasso and Post-Lasso model.
}
\description{
\emph{cv.plasso()} uses the \code{\link[glmnet]{glmnet}} package to estimate the coefficient paths and cross-validates least squares Lasso AND Post-Lasso.
}
