% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.plasso.R
\name{cv.plasso}
\alias{cv.plasso}
\title{Cross-validated Lasso and Post-Lasso}
\usage{
cv.plasso(x, y, w = NULL, kf = 10, parallel = FALSE, ...)
}
\arguments{
\item{x}{Matrix of covariates (number of observations times number of covariates matrix)}

\item{y}{Vector of outcomes}

\item{w}{Vector of weights}

\item{kf}{Number of folds in k-fold CV}

\item{parallel}{Set as FALSE for parallelized cross-validation}

\item{...}{Pass \code{\link{glmnet}} options}
}
\value{
List with the names of selected variables at cross-validated minima for Lasso and Post-Lasso.
}
\description{
\emph{cv.plasso()} uses the \code{\link{glmnet}} package to estimate the coefficient paths and cross-validates least squares Lasso AND Post-Lasso.
}
