% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasso.R
\name{CV_core}
\alias{CV_core}
\title{Core part for (Post-) Lasso cross-validation}
\usage{
CV_core(x, y, w, cvgroup, list, i, lambda, ...)
}
\arguments{
\item{x}{Covariate matrix to be used in CV}

\item{y}{Vector of outcomes}

\item{w}{Vector of weight}

\item{cvgroup}{Categorical with k groups to identify folds}

\item{list}{List 1:k}

\item{i}{Number of fold that is used for prediction}

\item{lambda}{Series of lambdas used}

\item{...}{Pass \code{\link{glmnet}} options}
}
\value{
MSE_lasso / MSE_plasso: means squared errors for each lambda.
}
\description{
\emph{CV_core()} ccontains the core parts of the CV for Lasso and Post-Lasso.
}
\keyword{internal}
