% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasso.R
\name{plasso}
\alias{plasso}
\title{Lasso and Post-Lasso}
\usage{
plasso(x, y, w = NULL, ...)
}
\arguments{
\item{x}{Matrix of covariates (number of observations times number of covariates matrix)}

\item{y}{Vector of outcomes}

\item{w}{Vector of weights}

\item{...}{Pass \code{\link[glmnet]{glmnet}} options}
}
\value{
List including glmnet object and names of selected variables at cross-validated minima for Lasso and Post-Lasso.
}
\description{
\emph{plasso()} implicitly estimates a Lasso model using the \code{\link[glmnet]{glmnet}} package
and additionally estimates coefficient paths for a subsequent Post-Lasso model.
}
