% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.plasso.R
\name{plot.cv.plasso}
\alias{plot.cv.plasso}
\title{Plot of cross-validation curves}
\usage{
\method{plot}{cv.plasso}(cv.plasso, lasso = FALSE)
}
\arguments{
\item{cv.plasso}{\code{\link{cv.plasso}} object}
}
\description{
Plot of cross-validation curves.
}
