% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.plasso.R
\name{predict.cv.plasso}
\alias{predict.cv.plasso}
\title{Predict after cross-validated (Post-) Lasso}
\usage{
\method{predict}{cv.plasso}(
  object,
  ...,
  newx = NULL,
  type = c("response", "coefficients"),
  s = c("optimal", "all"),
  se_rule = 0
)
}
\arguments{
\item{object}{Fitted \code{\link{cv.plasso}} model object}

\item{...}{Pass generic \code{\link[stats]{predict}} options}

\item{newx}{Matrix of new values for x at which predictions are to be made. If no value is supplied, x from fitting procedure is used. This argument is not used for \code{type="coefficients"}.}

\item{type}{Type of prediction required. \code{"response"} returns fitted values, \code{"coefficients"} returns beta estimates.}

\item{s}{Determines whether prediction is done for all values of lambda (\code{"all"}) or only for the optimal lambda (\code{"optimal"}) according to the standard error-rule.}

\item{se_rule}{If equal to 0, predictions from cross-validated MSE minimum (default). Negative values go in the direction of smaller
models, positive values go in the direction of larger models (e.g. \code{se_rule=-1} creates the standard 1SE rule).
This argument is not used for \code{s="all"}.}
}
\value{
Returns predictions of coefficients or fitted values for all lambda values or the optimal one.
}
\description{
Prediction for cross-validated (Post-) Lasso.
}
