% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasso.R
\name{print.plasso}
\alias{print.plasso}
\title{Print (Post-) Lasso model}
\usage{
\method{print}{plasso}(plasso, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{plasso}{\code{\link{plasso}} object}

\item{digits}{Integer, used for number formatting}

\item{...}{Pass generic R print options}
}
\value{
Prints cross-validated MSE and active variables for Lasso and Post-Lasso.
}
\description{
Printing main insights from (Post-) Lasso model.
}
