% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.plasso.R
\name{print.summary.cv.plasso}
\alias{print.summary.cv.plasso}
\title{Print summary of (Post-) Lasso model}
\usage{
\method{print}{summary.cv.plasso}(x, ..., digits = max(3L, getOption("digits") - 3L))
}
\arguments{
\item{x}{Summary of plasso object (either of class "summary.plasso' or "summaryDefault")}

\item{...}{Pass generic R \code{\link[base]{print}} options}

\item{digits}{Integer, used for number formatting}
}
\description{
Prints summary information of cv.plasso object
}
