% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.plasso.R
\name{print.summary.cv.plasso}
\alias{print.summary.cv.plasso}
\title{Print summary of (Post-) Lasso model}
\usage{
\method{print}{summary.cv.plasso}(object, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{Summary of plasso object (either of class "summary.plasso' or "summaryDefault")}

\item{digits}{Integer, used for number formatting}

\item{...}{Pass generic R print options}
}
\description{
Prints summary information of cv.plasso object
}
