% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.plasso.R
\name{summary.cv.plasso}
\alias{summary.cv.plasso}
\title{Summary of cross-validated (Post-) Lasso model}
\usage{
\method{summary}{cv.plasso}(object, ..., default = TRUE)
}
\arguments{
\item{object}{\code{\link[plasso]{cv.plasso}} object}

\item{...}{Pass generic \code{\link[base]{summary}} summary options}

\item{default}{TRUE for glmnet-like summary output, FALSE for more specific summary information}
}
\value{
For specific summary information: List object containing optimal lambda value and associated MSE for both cross-validated Lasso and Post-Lasso model.
}
\description{
Summary of cross-validated (Post-) Lasso model.
}
